+++ src/components/Calculator.jsx
@@ -1,12 +1,8 @@
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import Button from './Button';
 import data from '../data';
 import Screen from './Screen';
 import calculate from '../logic/calculate';
-import fecthQuotes from '../api';
-import Loading from './Loading';
-import Error from './Error';
-import Quotes from './Quotes';
 
 const Calculator = () => {
   const [calcObj, setCalcObj] = useState({
@@ -14,35 +10,12 @@ const Calculator = () => {
     next: null,
     operation: null,
   });
-  const [quote, setQuote] = useState(null);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState(null);
-  const retrieveData = async () => {
-    setLoading(true);
-    try {
-      const data = await fecthQuotes();
-      setQuote(data[0]);
-    } catch (error) {
-      setError(error);
-    } finally {
-      setLoading(false);
-    }
-  };
-  useEffect(() => {
-    retrieveData();
-  }, []);
   const handleClick = (buttonName) => {
     const newObj = calculate(calcObj, buttonName);
     setCalcObj(newObj);
   };
-  if (loading) {
-    return <Loading />;
-  }
-  if (error) {
-    return <Error message={error} />;
-  }
   return (
-    <div className=" w-[90%] md:w-[500px] mx-auto">
+    <div className=" w-[90%] md:w-[400px] mx-auto">
       <Screen value={calcObj.next || (calcObj.total || '0')} />
       <div className="grid grid-cols-4">
         {
@@ -56,7 +29,6 @@ const Calculator = () => {
         ))
       }
       </div>
-      {quote && <Quotes value={quote} />}
     </div>
   );
 };
