+++ src/components/Loading.jsx
@@ -1,7 +1,5 @@
-import React from 'react';
-
 const Loading = () => (
-  <div className=" text-4xl text-center text-white">Loading...</div>
+  <div className=" text-4xl text-center text-black mt-20">Loading...</div>
 );
 
 export default Loading;
